# filter the compared benchmark results by name using glob pattern(s)