@billwert 2019-07-23
> Looking forward to getting more regular results up on benchview.

Are you not seeing results? They should be there.