@adamsitnik 2019-07-24
It looks like no results were uploaded in the last two weeks.

Is the BenchView Uploader Service down again? Is the new Reporting System going to have more reliable uploader?